[DEV] update and refactor code
