+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_zoom","path":"/Users/webvedant/Documents/Flutter/Work/modifiedPackages/flutter_zoom/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_zoom","path":"/Users/webvedant/Documents/Flutter/Work/modifiedPackages/flutter_zoom/","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"flutter_zoom","path":"/Users/webvedant/Documents/Flutter/Work/modifiedPackages/flutter_zoom/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_zoom","dependencies":[]}],"date_created":"2022-12-15 17:00:29.835200","version":"3.3.4"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_zoom","path":"/Users/admin/Documents/source/flutter_zoom/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_zoom","path":"/Users/admin/Documents/source/flutter_zoom/","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"flutter_zoom","path":"/Users/admin/Documents/source/flutter_zoom/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_zoom","dependencies":[]}],"date_created":"2023-04-06 15:06:11.289470","version":"3.7.3"}
