+++ example/lib/main.dart
@@ -1,26 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_zoom_example/meeting_bindings.dart';
+import 'package:get/get.dart';
 import 'meeting_screen.dart';
 
-// for complete example see https://github.com/evilrat/flutter_zoom/tree/master/example
+void main() => runApp(const MyApp());
 
-void main() => runApp(const ExampleApp());
-
-class ExampleApp extends StatelessWidget {
-  const ExampleApp({Key? key}) : super(key: key);
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Example Zoom SDK',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      navigatorObservers: const [],
-      debugShowCheckedModeBanner: false,
-      initialRoute: '/',
-      routes: {
-        '/': (context) => const MeetingWidget(),
-      },
+    return GetMaterialApp(
+      title: 'zoom meeting',
+      defaultTransition: Transition.fadeIn,
+      home: const MeetingScreen(),
+      initialBinding: MeetingBindings(),
     );
   }
 }
