+++ example/lib/widgets/text_field.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class TextFieldWidget extends StatelessWidget {
+  const TextFieldWidget({Key? key, required this.controller, required this.label}) : super(key: key);
+  final TextEditingController? controller;
+  final String? label;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8.0),
+      child: TextField(
+        controller: controller,
+        decoration: InputDecoration(
+          border: const OutlineInputBorder(),
+          labelText: label,
+        ),
+      ),
+    );
+  }
+}
