+++ example/pubspec.yaml
@@ -23,6 +23,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  get: ^4.6.5
 
 dev_dependencies:
   flutter_lints: ^1.0.4
