+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: flutter_zoom
-description: A new Flutter project.
+description: Flutter with zoom plugin
 version: 0.0.1
 homepage:
